Migrate shared CA code into includable module
