+++ lib/terrafying/components/letsencrypt.rb
@@ -1,4 +1,5 @@
 
+require 'terrafying/components/ca'
 require 'terrafying/generator'
 
 module Terrafying
@@ -8,6 +9,8 @@ module Terrafying
     class LetsEncrypt < Terrafying::Context
 
       attr_reader :name, :source
+
+      include CA
 
       def self.create(bucket, options={})
         LetsEncrypt.new.create bucket, options
@@ -57,10 +60,6 @@ module Terrafying
         @source = File.join("s3://", @bucket, @prefix, @name, "ca.cert")
 
         self
-      end
-
-      def create_keypair(name, options={})
-        create_keypair_in(self, name, options)
       end
 
       def create_keypair_in(ctx, name, options={})
@@ -118,36 +117,6 @@ module Terrafying
                      }
 
         reference_keypair(name)
-      end
-
-      def reference_keypair(name)
-        key_ident = "#{@name}-#{name.gsub(/\./, '-')}"
-
-        {
-          name: name,
-          ca: self,
-          source: {
-            cert: File.join("s3://", @bucket, @prefix, @name, name, "cert"),
-            key: File.join("s3://", @bucket, @prefix, @name, name, "key"),
-          },
-          resources: [ "aws_s3_bucket_object.#{key_ident}-key", "aws_s3_bucket_object.#{key_ident}-cert" ],
-          iam_statement: {
-            Effect: "Allow",
-            Action: [
-              "s3:GetObjectAcl",
-              "s3:GetObject",
-            ],
-            Resource: [
-              "arn:aws:s3:::#{File.join(@bucket, @prefix, @name, "ca.cert")}",
-              "arn:aws:s3:::#{File.join(@bucket, @prefix, @name, name, "cert")}",
-              "arn:aws:s3:::#{File.join(@bucket, @prefix, @name, name, "key")}",
-            ]
-          }
-        }
-      end
-
-      def <=>(other)
-        @name <=> other.name
       end
 
     end
