+++ lib/terrafying/components/selfsignedca.rb
@@ -1,4 +1,5 @@
 
+require 'terrafying/components/ca'
 require 'terrafying/generator'
 
 module Terrafying
@@ -8,6 +9,8 @@ module Terrafying
     class SelfSignedCA < Terrafying::Context
 
       attr_reader :name, :source, :ca_key
+
+      include CA
 
       def self.create(name, bucket, options={})
         SelfSignedCA.new.create name, bucket, options
@@ -64,9 +67,6 @@ module Terrafying
         self
       end
 
-      def create_keypair(name, options={})
-        create_keypair_in(self, name, options)
-      end
 
       def create_keypair_in(ctx, name, options={})
         options = {
@@ -122,36 +122,6 @@ module Terrafying
                      }
 
         reference_keypair(name)
-      end
-
-      def reference_keypair(name)
-        key_ident = "#{@name}-#{name.gsub(/\./, '-')}"
-
-        {
-          name: name,
-          ca: self,
-          source: {
-            cert: File.join("s3://", @bucket, @prefix, @name, name, "cert"),
-            key: File.join("s3://", @bucket, @prefix, @name, name, "key"),
-          },
-          resources: [ "aws_s3_bucket_object.#{key_ident}-key", "aws_s3_bucket_object.#{key_ident}-cert" ],
-          iam_statement: {
-            Effect: "Allow",
-            Action: [
-              "s3:GetObjectAcl",
-              "s3:GetObject",
-            ],
-            Resource: [
-              "arn:aws:s3:::#{File.join(@bucket, @prefix, @name, "ca.cert")}",
-              "arn:aws:s3:::#{File.join(@bucket, @prefix, @name, name, "cert")}",
-              "arn:aws:s3:::#{File.join(@bucket, @prefix, @name, name, "key")}",
-            ]
-          }
-        }
-      end
-
-      def <=>(other)
-        @name <=> other.name
       end
 
     end
